# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CommandAccepted": {
        "description": "What every control route answers.\n\nOne shape for pause, resume, cancel, checkpoint and patch, because a client\nscripting them wants one thing to read: what the run is now, and when what it\nasked for lands.",
        "properties": {
          "applied": {
            "description": "The dotted paths a `PATCH` accepted, in a fixed order. Empty elsewhere.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "applies_at_iteration": {
            "description": "The iteration the request takes effect at - the next one, since a run is\nonly interruptible at its progress callback. Absent when the effect\nwas immediate, as for a run cancelled before it ever started.",
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "type": "object"
      },
...
      "ConfigDocument": {
        "additionalProperties": false,
        "description": "The configuration *document*: the exact schema of the CLI's TOML file.\n\nPublic because it is the one schema three frontends share: the CLI reads it\nfrom TOML, the server accepts it as JSON, and the resolver renders its own\noutput back through it. The resolver builds one of these instead of a\n[`RunConfig`](crate::RunConfig) directly, so that \"the resolver cannot\nproduce a config the CLI would refuse\" holds *by construction*: the only way\nfrom a document to a `RunConfig` is [`build`](crate::build()), which is what\n`load` calls.\n\n[`RunConfig`](crate::RunConfig), by contrast, stays internal: it is the\nengine's shape and changes at the engine's pace.",
        "properties": {
          "agent": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentToml"
              }
            ],
            "x-retrograd-applies-to": [
              "agent_grpo"
            ]
          },
          "checkpoint": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CheckpointToml"
              }
            ]
          },
          "distill": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DistillToml"
              }
            ],
            "x-retrograd-applies-to": [
              "distill"
            ]
          },
          "evaluation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EvaluationToml"
              }
            ]
          },
          "grpo": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GrpoToml"
              }
            ],
            "x-retrograd-applies-to": [
              "grpo"
            ]
          },
          "lora": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LoraToml",
                "description": "`[lora]`: required by `lora` and `hybrid`, refused by `full` and\n`partial`. Optional as a *document* section because a run that trains\nbase tensors has no adapter at all, and demanding rank, alpha and\ntargets from it would be demanding the shape of something it does not\ncreate."
              }
            ]
          },
          "metrics": {
            "$ref": "#/components/schemas/MetricsToml"
          },
          "model": {
            "$ref": "#/components/schemas/ModelToml"
          },
          "observe": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ObserveToml"
              }
            ],
            "x-retrograd-applies-to": [
              "ppo",
              "grpo",
              "agent_grpo"
            ]
          },
          "optimizer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OptimizerToml",
                "description": "`[optimizer.<name>]`: the chosen optimizer's own knobs. A table for an\noptimizer the run did not choose is refused at build time."
              }
            ]
          },
          "output": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OutputToml",
                "description": "`[output]`: where the run's result goes, and which kind of result it is.\nThe kind is defaulted from the policy when absent; the path is not, so\na document that names no output is refused."
              }
            ]
          },
          "ppo": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PpoToml"
              }
            ],
            "x-retrograd-applies-to": [
              "ppo"
            ]
          },
          "preference": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PreferenceToml"
              }
            ],
            "x-retrograd-applies-to": [
              "preference"
            ]
          },
          "reference": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReferenceToml",
                "description": "`[reference]`: the frozen model a KL penalty is scored against. Its own\nsection rather than a field of `[ppo]` or `[grpo]`, because one file\nserves every objective that declares a KL term."
              }
            ]
          },
          "run": {
            "$ref": "#/components/schemas/RunToml"
          },
          "sft": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SftToml"
              }
            ],
            "x-retrograd-applies-to": [
              "sft"
            ]
          },
          "trainable": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TrainableToml",
                "description": "`[trainable]`: which base tensors a `partial` or `hybrid` run selects.\nAbsent for `lora`, and refused when present alongside it - a selector a\npolicy ignores is a selector the user believes is in effect."
              }
            ]
          },
          "training": {
            "$ref": "#/components/schemas/TrainingToml"
          }
        },
        "required": [
          "run"
        ],
        "type": "object"
      },
...
      "ForkFrom": {
        "additionalProperties": false,
        "description": "Where a fork starts from.\n\nTwo shapes, and the difference is who supplies the configuration. `run` names\na run this server knows, so the parent's own effective configuration is\nreused - that is the fork, and the resume. `path` names a checkpoint\ndirectory, which carries a manifest and no configuration at all, so a `recipe`\nor a `config` has to come with it.",
        "properties": {
          "checkpoint": {
            "description": "A checkpoint id under the parent's checkpoint directory (`step-400`,\n`best`), or absent for the most recent one - the same resolution\n`--resume` performs.",
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "description": "A `.state` directory, for a checkpoint written by something other than\nthis server.",
            "type": [
              "string",
              "null"
            ]
          },
          "run": {
            "description": "The parent run's id.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "PatchRequest": {
        "additionalProperties": false,
        "description": "`PATCH /v1/runs/{id}` - the whitelist of the contract, and nothing else.\n\n`deny_unknown_fields` is what makes the list closed: `n_ctx`, `lora`,\n`algorithm` and every path are rejected by the deserializer before a handler\nruns, and the handler turns that into the 409 that points at `fork_from`.\nField names follow the *document* grammar (`training.lr`), the same one\n`params`, `provenance` and `GET /v1/defaults` speak.",
        "properties": {
          "checkpoint": {
            "default": null,
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CheckpointPatch"
              }
            ]
          },
          "evaluation": {
            "default": null,
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EvaluationPatch"
              }
            ]
          },
          "training": {
            "default": null,
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TrainingPatch"
              }
            ]
          }
        },
        "type": "object"
      },
...
      "PlanRequest": {
        "additionalProperties": false,
        "description": "`POST /v1/plan` and `POST /v1/runs`.\n\nExactly one of `recipe` (form (a): an intention) or `config` (form (b): a\ncomplete configuration, the same schema as the CLI's TOML). Form (c) - a raw\nTOML body - arrives as `config` too, parsed by the extractor.",
        "properties": {
          "config": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConfigDocument",
                "description": "A whole configuration document. Skips the semantic phases but not the\nbudget check."
              }
            ]
          },
          "fork_from": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ForkFrom",
                "description": "Continue from a checkpoint. With `fork_from.run` this is the whole\nbody: the parent supplies the configuration and `params` go on top."
              }
            ]
          },
          "name": {
            "description": "Human label. No uniqueness is imposed; identity is the run id.",
            "type": [
              "string",
              "null"
            ]
          },
          "params": {
            "description": "The client's own parameters: a partial document tree, spelled exactly as\nthe CLI's TOML spells it. Every leaf it sets is **locked** - no phase may\nre-derive it - and everything it leaves out the server derives, from the\nrules `GET /v1/defaults` publishes.",
            "type": "object"
          },
          "recipe": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ]
          }
        },
        "type": "object"
      },
      "PlanResponse": {
        "description": "What a resolution renders back.",
        "properties": {
          "effective_config": {
            "$ref": "#/components/schemas/ConfigDocument",
            "description": "The effective configuration, in the same schema `config` takes on input.\nServer-declared values are redacted to their catalogue id."
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "plan": {
            "$ref": "#/components/schemas/PlanSummary"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance",
            "description": "Flat map, keyed by dotted field path."
          }
        },
        "required": [
          "effective_config",
          "provenance",
          "plan"
        ],
        "type": "object"
      },
...
      "Problem": {
        "description": "Wire shape of a problem document (RFC 9457). Serialized by hand rather than\nderived from `ApiError` so the field order - and therefore the bytes - is\nfixed. Published as the `Problem` component: every failure of the API but\nthe OpenAI routes' has this body.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "meta": {
            "description": "The payload specific to the problem type, e.g. the dominant memory\nposts of `insufficient-memory`.",
            "type": [
              "object",
              "null"
            ]
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "description": "The id every log line of this request carries.",
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "description": "`https://retrograd.dev/problems/<slug>`: the slug is what a client\nbranches on.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
      "Provenance": {
        "additionalProperties": {
          "$ref": "#/components/schemas/Origin"
        },
        "description": "The provenance map. `BTreeMap` so the serialized bytes are the same for two\nidentical resolutions (invariant 3).",
        "propertyNames": {
          "type": "string"
        },
        "type": "object"
      },
...
      "RunListing": {
        "properties": {
          "next_cursor": {
            "description": "Opaque: pass it back as `?cursor=` for the next page. Absent on the last\npage, which is how a client knows to stop.",
            "type": [
              "string",
              "null"
            ]
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array"
          }
        },
        "required": [
          "runs"
        ],
        "type": "object"
      },
...
      "RunView": {
        "description": "`GET /v1/runs/{id}` and the `201` of `POST /v1/runs`.\n\nThe three heavy fields are carried as pre-rendered JSON. They are produced by\nserializing the typed resolution once, at creation, and then written to the\njournal and served verbatim - which is both cheaper than re-rendering and the\nonly way a run restored from disk answers the same bytes it answered while it\nwas alive.",
        "properties": {
          "algorithm": {
            "description": "The `run.algorithm` spelling.",
            "type": "string"
          },
          "created_at": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "effective_config": {
            "$ref": "#/components/schemas/ConfigDocument"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "holds_device": {
            "description": "Weights, KV caches and optimizer state stay allocated while a run is\npaused. True whenever this run is holding the device.",
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "model": {
            "description": "The base model's file name.",
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "objective": {
            "type": [
              "string",
              "null"
            ]
          },
          "observed": {
            "type": "boolean"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanSummary"
          },
          "progress": {
            "$ref": "#/components/schemas/RunProgress"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "queue_position": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "started_at": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "id",
          "status",
          "algorithm",
          "model",
          "observed",
          "created_at",
          "progress",
          "holds_device",
          "effective_config",
          "provenance",
          "plan"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP control plane for LoRA training runs.",
    "title": "retrograd-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/plan": {
      "post": {
        "operationId": "post_plan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Resolve a recipe without creating anything"
      }
    },
...
    "/v1/runs": {
      "get": {
        "operationId": "get_runs",
        "parameters": [
          {
            "description": "Only runs in this state.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only runs of this `run.algorithm` (sft, grpo, ppo, preference, …).",
            "in": "query",
            "name": "algorithm",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only runs with exactly this name.",
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page size, 1 to 500. Default 50.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "The id of the last run of the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListing"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "List runs, newest first"
      },
      "post": {
        "operationId": "post_runs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "a dry run, or a repeat of an Idempotency-Key: no run was started"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "created"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Create a run"
      }
    },
    "/v1/runs/{id}": {
      "delete": {
        "operationId": "delete_runs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "deleted"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Forget a finished run"
      },
      "get": {
        "operationId": "get_runs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "One run: state, progress, effective config, plan"
      },
      "patch": {
        "operationId": "patch_runs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAccepted"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Adjust a schedule on a live run"
      }
    },
...
  }
}